# ai-recruiting-platform/packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI Recruiting Platform API
  version: 1.0.0
  description: Production-grade AI recruiting platform with automated sourcing, scoring, and outreach
servers:
  - url: https://api.recruiting-platform.com/v1
    description: Production server
  - url: https://staging-api.recruiting-platform.com/v1
    description: Staging server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /webhooks/outreach:
    post:
      summary: Webhook for outreach events
      operationId: webhooks.outreach
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookEvent'
      responses:
        '200':
          description: Webhook processed successfully
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    WebhookEvent:
      type: object
      required:
        - type
        - timestamp
        - data
      properties:
        type:
          type: string
          enum: [outreach.opened, outreach.replied, outreach.bounced, calendar.created, calendar.updated, ats.sync]
        timestamp:
          type: string
          format: date-time
        data:
          type: object
        signature:
          type: string
          description: HMAC signature for verification
...